Separate install command into its own file
